requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
taskipy
boto3
paramiko
